refactor: extract internal event (click tree item)
